+++ translations/VideoBundle.de.php
@@ -65,7 +65,7 @@ return array (
   'Sort by' => 'Sortieren',
   'Date' => 'Datum',
   'Date desc' => 'Datum absteigend',
-  '' => '',
+  'Video group statistics' => 'Videogruppen-Statistiken',
   '' => '',
   '' => '',
   '' => '',